requirements.txt
fastapi
uvicorn
python-multipart
langchain
langchain-google-genai==4.2.1
langchain-community
langchain-chroma
langchain-text-splitters
gitpython
chromadb
fastembed
python-dotenv
networkx
tree-sitter
tree-sitter-python
tree-sitter-javascript
tree-sitter-java
pydantic
asyncio
aiohttp
python-dotenv
